it now works also under submodule directory
