+++ lib/git_stats/base.rb
@@ -8,13 +8,12 @@ require 'git_stats/version'
 require 'git_stats/i18n'
 require 'git_stats/cli'
 require 'git_stats/generator'
-require 'git_stats/validator'
+require 'git_stats/command_runner'
+require 'git_stats/command_parser'
 
 require 'git_stats/git_data/activity'
 require 'git_stats/git_data/author'
 require 'git_stats/git_data/blob'
-require 'git_stats/git_data/command_parser'
-require 'git_stats/git_data/command_runner'
 require 'git_stats/git_data/commit'
 require 'git_stats/git_data/repo'
 require 'git_stats/git_data/short_stat'
