--- lib/git_stats/git_data/command_runner.rb
@@ -1,15 +1,0 @@
-# -*- encoding : utf-8 -*-
-module GitStats
-  module GitData
-    class CommandRunner
-      def run(path, command)
-        execute(command, path).encode!('UTF-8', 'UTF-8', :invalid => :replace)
-      end
-
-      private
-      def execute(command, path)
-        Dir.chdir(path) { %x[#{command}] }
-      end
-    end
-  end
-end
